documentation 5/5

+++ src/nuclear_data.rs
@@ -31,6 +31,7 @@ pub struct NuclearDataReaction<T: Float> {
 }
 
 impl<T: Float> NuclearDataReaction<T> {
+    /// Constructor.
     pub fn new(
         rtype: ReactionType,
         nu_bar: T,
@@ -41,10 +42,17 @@ impl<T: Float> NuclearDataReaction<T> {
         todo!()
     }
 
+    /// Get the cross section for the specified group.
     pub fn get_cross_section(group: usize) -> T {
         todo!()
     }
 
+    /// Uses RNG to get new angle and energy after a reaction. This 
+    /// needs to be Rustified right away as this is too C-like. 
+    /// The current function takes to pointer as arguments that act as
+    /// arrays (energy/angle_out), a max length for those arrays (max_...) 
+    /// and a pointer to an integer that gets updated with the actual 
+    /// length of the arrays (n_out). Change this asap
     pub fn sample_collision(
         incident_energy: T,
         material_mass: T,
@@ -58,14 +66,15 @@ impl<T: Float> NuclearDataReaction<T> {
     }
 }
 
-/// Structure used to hold a list of reactions. Not defined
-/// via an alias because we need an impl block.
+/// Structure used to hold a list of reactions.
 #[derive(Debug)]
 pub struct NuclearDataSpecies<T: Float> {
+    /// List of reactions
     pub reactions: Vec<NuclearDataReaction<T>>,
 }
 
 impl<T: Float> NuclearDataSpecies<T> {
+    /// Adds a reaction to the internal list.
     pub fn add_reaction(
         &mut self,
         rtype: ReactionType,
@@ -85,16 +94,22 @@ pub type NuclearDataIsotope<T> = Vec<NuclearDataSpecies<T>>;
 /// nuclear data.
 #[derive(Debug)]
 pub struct NuclearData<T: Float> {
+    /// Total number of energy groups?
     pub num_energy_groups: u32, //usize?
+    /// Reactions and cross sections are stored by isotopes,
+    /// those being stored by species
     pub isotopes: Vec<NuclearDataIsotope<T>>,
+    /// Overall energy layout
     pub energies: Vec<T>,
 }
 
 impl<T: Float> NuclearData<T> {
+    /// Constructor.
     pub fn new(num_groups: u32, energy_low: T, energy_high: T) -> Self {
         todo!()
     }
 
+    /// Adds an isotope to the internal list.
     #[allow(clippy::too_many_arguments)]
     pub fn add_isotope(
         &mut self,
@@ -111,20 +126,24 @@ impl<T: Float> NuclearData<T> {
         todo!()
     }
 
+    /// Returns the energy group a specific energy belongs to.
     pub fn get_energy_groups(&self, energy: T) -> u32 {
         //usize?
         todo!()
     }
 
-    pub fn get_number_reactions(&self, isotope_index: usize) -> u32 {
+    /// Returns the number of reactions for a given isotope.
+    pub fn get_number_reactions(&self, isotope_index: usize) -> usize {
         //usize?
         todo!()
     }
 
+    /// Returns the total cross section for a given energy group.
     pub fn get_total_cross_section(&self, isotope_index: usize, group: usize) -> T {
         todo!()
     }
 
+    /// Returns the reaction-specific cross section for a given energy_group.
     pub fn get_reaction_cross_section(
         &self,
         react_index: usize,
